API/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
numpy
opencv-python
supervision
inference-sdk
google-generativeai
python-multipart
python-dotenv
